Refactor: change all request and response represent class to DTO class. (#7)
