+++ src/main/java/com/dtvn/springbootproject/dto/responseDtos/Account/AccountResponseDTO.java
@@ -1,4 +1,4 @@
-package com.dtvn.springbootproject.requests;
+package com.dtvn.springbootproject.dto.responseDtos.Account;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,11 +9,11 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class RegisterAnAccountRequest {
+public class AccountResponseDTO {
+    private int account_id;
     private String firstname;
     private String lastname;
     private String email;
-    private String password;
     private String role;
     private String address;
     private String phone;
