+++ src/main/java/com/dtvn/springbootproject/services/implementations/AuthenticationServiceImpl.java
@@ -1,8 +1,8 @@
 package com.dtvn.springbootproject.services.implementations;
 
 //import com.dtvn.springbootproject.exceptions.AuthenticationException;
-import com.dtvn.springbootproject.requests.AuthenticationRequest;
-import com.dtvn.springbootproject.responses.AuthenticationResponse;
+import com.dtvn.springbootproject.dto.requestDtos.Auth.AuthenticationRequestDTO;
+import com.dtvn.springbootproject.dto.responseDtos.Auth.AuthenticationResponseDTO;
 import com.dtvn.springbootproject.repositories.AccountRepository;
 import com.dtvn.springbootproject.config.JwtService;
 import com.dtvn.springbootproject.services.interfaces.AuthenticationService;
@@ -23,7 +23,7 @@ public class AuthenticationServiceImpl implements AuthenticationService {
     private final AuthenticationManager authenticationManager;
 
     @Override
-    public AuthenticationResponse login(AuthenticationRequest request) {
+    public AuthenticationResponseDTO login(AuthenticationRequestDTO request) {
         try {
             authenticationManager.authenticate(
                     new UsernamePasswordAuthenticationToken(
@@ -32,7 +32,7 @@ public class AuthenticationServiceImpl implements AuthenticationService {
                     )
             );
         } catch (AuthenticationException e) {
-            return AuthenticationResponse.builder()
+            return AuthenticationResponseDTO.builder()
                     .code(403)
                     .message(e.getMessage())
                     .build();
@@ -41,7 +41,7 @@ public class AuthenticationServiceImpl implements AuthenticationService {
                 .orElseThrow();
         var jwtToken = jwtService.generateToken(account);
         var refreshToken = jwtService.generateRefreshToken(account);
-        return AuthenticationResponse.builder()
+        return AuthenticationResponseDTO.builder()
                 .code(200)
                 .message("Login successfully")
                 .access_token(jwtToken)
