+++ src/main/java/com/dtvn/springbootproject/services/interfaces/AccountService.java
@@ -1,8 +1,8 @@
 package com.dtvn.springbootproject.services.interfaces;
 
-import com.dtvn.springbootproject.requests.RegisterAnAccountRequest;
-import com.dtvn.springbootproject.responses.AccountResponse;
+import com.dtvn.springbootproject.dto.requestDtos.Account.AccountRegisterRequestDTO;
+import com.dtvn.springbootproject.dto.responseDtos.Account.AccountResponseDTO;
 
 public interface AccountService {
-    AccountResponse registerAnAccount(RegisterAnAccountRequest request);
+    AccountResponseDTO registerAnAccount(AccountRegisterRequestDTO request);
 }
